refactor: create dedicated progress feature
- Create new progress feature folder structure for better organization
- Move ProgressWithFeedback component from UI to progress/components
- Move useProgressFeedback hook from general hooks to progress/hooks
- Create barrel files for cleaner imports and exports
- Update import paths in affected files
- Fix import path for shareStatements in ShareEmailModal
