+++ src/features/progress/components/index.ts
@@ -1,0 +1,1 @@
+export { default as ProgressWithFeedback } from './ProgressWithFeedback';
