+++ src/features/progress/index.ts
@@ -1,0 +1,9 @@
+/**
+ * Progress Feature Module
+ * 
+ * Provides functionality for tracking and displaying user progress
+ * throughout the application.
+ */
+
+export { useProgressFeedback } from './hooks';
+export { ProgressWithFeedback } from './components';
